sztu-v5-final/requirements.txt
# ============================================================
# 迈克尔逊干涉测杨氏模量一体化实验系统
# requirements.txt - Vercel 云端部署依赖清单
# Python 3.10+
# ============================================================

# ── Web 框架 ─────────────────────────────────
streamlit>=1.35,<2.0

# ── 计算机视觉（P0 计数模块核心依赖）────────
opencv-python-headless>=4.10,<5.0   # 云端无 GUI，必须用 headless 版
numpy>=1.26,<2.0

# ── 科学计算（P1 不确定度计算）────────────────
scipy>=1.12,<1.14

# ── 数据可视化 ────────────────────────────────
matplotlib>=3.8,<3.9
pandas>=2.2,<2.3

# ── AI / LLM（P2 智能体 + P3 Vision 诊断）────
anthropic>=0.28,<1.0
openai>=1.40,<2.0

# ── PDF 报告导出（P1 可选）────────────────────
fpdf2>=2.7.9,<3.0

# ── 深度学习推理（P0 CNN 层，可选升级）────────
# onnxruntime>=1.18.0     # 取消注释以启用 ONNX 模型推理
# 注：onnxruntime 包体较大（~200MB），Vercel 限制 250MB，按需启用

# ── QR Code 生成（分享二维码）────────────────
qrcode[pil]>=7.4.2,<8.0
Pillow>=10.3,<11.0

# ── 工具库 ────────────────────────────────────
python-dotenv>=1.0,<2.0      # 本地开发 .env 环境变量
requests>=2.32,<3.0
